Version 2.3
Mostly tweaks, adjustments, and bug-fixes, but also a new data output system and additions to the custom settings options!

The program now outputs a csv file with stats for the games simulated, which can be opened in a spreadsheet. (I've only tested it on Google Sheets, but it should work on Excel etc. as well)

Added a ball-by-ball record of high scoring overs.

Hattricks taking place in a single over are now noted in the game log. (Although at the moment run outs are also counted as part of hattricks, which they shouldn't be)

Added option in settings.txt to control the toss and which team bats first.

Added option in settings.txt to pick captains/wicket-keepers.

Added option in settings.txt to show bowling changes in the game log.

Added description to customteams.txt of how player data lines are constructed.

Made on-going improvements to the match report.

Made captains more conservative in choosing to enforce the follow-on.

Made captains willing to sometimes declare if they are in the 90s.

Made adjustments and improvements to the bowler selection algorithm.

Made adjustments to 3rd and 4th innings batting aggression.

Made lots of adjustments and improvements to the altcricket selection process.

Fixed issue where wides would not result in an extra ball in the over.

Fixed issue where only four runs were awarded, instead of five, on wides that went to the boundary.

Fixed issue with custom/historical automatic team selection that made teams too likely to have five specialist bowlers.

Fixed issue with altcricket where Pakistan would not always be treated as the home team for series in the UAE.

Fixed issue with altcricket where recovering from injury would reactivate players who shouldn't have been avaliable.

Fixed issue in altcricker where the Man of the Match in a series with one game would not always also be the Man of the Series.

Fixed issue with header for custom team scorecards.

Fixed issue where, if a game was won by the team batting fourth with the very last possible ball of the match, the result would be incorrectly identified.